LPD-27759 Create new classes to reuse code
