+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-test/src/testIntegration/java/com/liferay/dynamic/data/mapping/upgrade/v5_1_1/test/DDMStructureDDMValidationUpgradeProcessTest.java
@@ -6,36 +6,10 @@
 package com.liferay.dynamic.data.mapping.upgrade.v5_1_1.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
-import com.liferay.change.tracking.test.util.BaseCTUpgradeProcessTestCase;
-import com.liferay.dynamic.data.mapping.form.field.type.constants.DDMFormFieldTypeConstants;
-import com.liferay.dynamic.data.mapping.io.DDMFormSerializer;
 import com.liferay.dynamic.data.mapping.model.DDMForm;
-import com.liferay.dynamic.data.mapping.model.DDMFormField;
-import com.liferay.dynamic.data.mapping.model.DDMFormFieldValidation;
-import com.liferay.dynamic.data.mapping.model.DDMFormFieldValidationExpression;
-import com.liferay.dynamic.data.mapping.model.DDMFormInstance;
-import com.liferay.dynamic.data.mapping.model.DDMStructure;
-import com.liferay.dynamic.data.mapping.service.DDMStructureLocalService;
-import com.liferay.dynamic.data.mapping.test.util.DDMFormTestUtil;
-import com.liferay.dynamic.data.mapping.test.util.DDMStructureTestUtil;
-import com.liferay.dynamic.data.mapping.util.DDMFormSerializeUtil;
-import com.liferay.portal.kernel.model.Group;
-import com.liferay.portal.kernel.model.change.tracking.CTModel;
-import com.liferay.portal.kernel.service.change.tracking.CTService;
-import com.liferay.portal.kernel.test.rule.AggregateTestRule;
-import com.liferay.portal.kernel.test.rule.DeleteAfterTestRun;
-import com.liferay.portal.kernel.test.util.GroupTestUtil;
-import com.liferay.portal.kernel.test.util.RandomTestUtil;
-import com.liferay.portal.kernel.upgrade.UpgradeProcess;
-import com.liferay.portal.test.rule.Inject;
-import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
-import com.liferay.portal.test.rule.PermissionCheckerMethodTestRule;
-import com.liferay.portal.upgrade.registry.UpgradeStepRegistrator;
-import com.liferay.portal.upgrade.test.util.UpgradeTestUtil;
+import com.liferay.dynamic.data.mapping.upgrade.BaseDDMStructureCTUpgradeProcessTestCase;
+import com.liferay.dynamic.data.mapping.upgrade.v5_1_1.test.util.DDMValidationUpgradeProcessTestUtil;
 
-import org.junit.Before;
-import org.junit.ClassRule;
-import org.junit.Rule;
 import org.junit.runner.RunWith;
 
 /**
@@ -43,99 +17,16 @@ import org.junit.runner.RunWith;
  */
 @RunWith(Arquillian.class)
 public class DDMStructureDDMValidationUpgradeProcessTest
-	extends BaseCTUpgradeProcessTestCase {
+	extends BaseDDMStructureCTUpgradeProcessTestCase {
 
-	@ClassRule
-	@Rule
-	public static final AggregateTestRule aggregateTestRule =
-		new AggregateTestRule(
-			new LiferayIntegrationTestRule(),
-			PermissionCheckerMethodTestRule.INSTANCE);
-
-	@Before
-	public void setUp() throws Exception {
-		_group = GroupTestUtil.addGroup();
+	@Override
+	protected String getClassName() {
+		return DDMValidationUpgradeProcessTestUtil.getClassName();
 	}
 
 	@Override
-	protected CTModel<?> addCTModel() throws Exception {
-		_ddmStructure = DDMStructureTestUtil.addStructure(
-			_group.getGroupId(), DDMFormInstance.class.getName(),
-			DDMFormTestUtil.createDDMForm(RandomTestUtil.randomString()));
-
-		setDefinition(_ddmStructure);
-
-		return _ddmStructureLocalService.updateDDMStructure(_ddmStructure);
+	protected DDMForm getDDMForm() {
+		return DDMValidationUpgradeProcessTestUtil.getDDMForm();
 	}
-
-	@Override
-	protected CTService<?> getCTService() {
-		return _ddmStructureLocalService;
-	}
-
-	@Override
-	protected void runUpgrade() throws Exception {
-		UpgradeProcess upgradeProcess = UpgradeTestUtil.getUpgradeStep(
-			_upgradeStepRegistrator, _CLASS_NAME);
-
-		upgradeProcess.upgrade();
-	}
-
-	protected void setDefinition(DDMStructure ddmStructure) {
-		DDMForm ddmForm = DDMFormTestUtil.createDDMForm();
-
-		DDMFormField ddmFormField = DDMFormTestUtil.createDDMFormField(
-			RandomTestUtil.randomString(), RandomTestUtil.randomString(),
-			DDMFormFieldTypeConstants.DATE, DDMFormFieldTypeConstants.DATE,
-			false, false, false);
-
-		DDMFormFieldValidation ddmFormFieldValidation =
-			new DDMFormFieldValidation();
-
-		ddmFormFieldValidation.setDDMFormFieldValidationExpression(
-			new DDMFormFieldValidationExpression() {
-				{
-					setName("dateRange");
-					setValue("dateRange({parameter} AND {parameter})");
-				}
-			});
-
-		ddmFormField.setDDMFormFieldValidation(ddmFormFieldValidation);
-
-		ddmForm.addDDMFormField(ddmFormField);
-
-		ddmStructure.setDefinition(
-			DDMFormSerializeUtil.serialize(ddmForm, _jsonDDMFormSerializer));
-	}
-
-	@Override
-	protected CTModel<?> updateCTModel(CTModel<?> ctModel) throws Exception {
-		_ddmStructure = (DDMStructure)ctModel;
-
-		setDefinition(_ddmStructure);
-
-		return _ddmStructureLocalService.updateDDMStructure(_ddmStructure);
-	}
-
-	private static final String _CLASS_NAME =
-		"com.liferay.dynamic.data.mapping.internal.upgrade.v5_1_1." +
-			"DDMValidationUpgradeProcess";
-
-	@Inject(
-		filter = "(&(component.name=com.liferay.dynamic.data.mapping.internal.upgrade.registry.DDMServiceUpgradeStepRegistrator))"
-	)
-	private static UpgradeStepRegistrator _upgradeStepRegistrator;
-
-	@DeleteAfterTestRun
-	private DDMStructure _ddmStructure;
-
-	@Inject
-	private DDMStructureLocalService _ddmStructureLocalService;
-
-	@DeleteAfterTestRun
-	private Group _group;
-
-	@Inject(filter = "ddm.form.serializer.type=json")
-	private DDMFormSerializer _jsonDDMFormSerializer;
 
 }
