+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-test/src/testIntegration/java/com/liferay/dynamic/data/mapping/upgrade/v5_1_1/test/util/DDMValidationUpgradeProcessTestUtil.java
@@ -1,0 +1,52 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2024 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+package com.liferay.dynamic.data.mapping.upgrade.v5_1_1.test.util;
+
+import com.liferay.dynamic.data.mapping.form.field.type.constants.DDMFormFieldTypeConstants;
+import com.liferay.dynamic.data.mapping.model.DDMForm;
+import com.liferay.dynamic.data.mapping.model.DDMFormField;
+import com.liferay.dynamic.data.mapping.model.DDMFormFieldValidation;
+import com.liferay.dynamic.data.mapping.model.DDMFormFieldValidationExpression;
+import com.liferay.dynamic.data.mapping.test.util.DDMFormTestUtil;
+import com.liferay.portal.kernel.test.util.RandomTestUtil;
+
+/**
+ * @author Carolina Barbosa
+ */
+public class DDMValidationUpgradeProcessTestUtil {
+
+	public static String getClassName() {
+		return "com.liferay.dynamic.data.mapping.internal.upgrade.v5_1_1." +
+			"DDMValidationUpgradeProcess";
+	}
+
+	public static DDMForm getDDMForm() {
+		DDMForm ddmForm = DDMFormTestUtil.createDDMForm();
+
+		DDMFormField ddmFormField = DDMFormTestUtil.createDDMFormField(
+			RandomTestUtil.randomString(), RandomTestUtil.randomString(),
+			DDMFormFieldTypeConstants.DATE, DDMFormFieldTypeConstants.DATE,
+			false, false, false);
+
+		DDMFormFieldValidation ddmFormFieldValidation =
+			new DDMFormFieldValidation();
+
+		ddmFormFieldValidation.setDDMFormFieldValidationExpression(
+			new DDMFormFieldValidationExpression() {
+				{
+					setName("dateRange");
+					setValue("dateRange({parameter} AND {parameter})");
+				}
+			});
+
+		ddmFormField.setDDMFormFieldValidation(ddmFormFieldValidation);
+
+		ddmForm.addDDMFormField(ddmFormField);
+
+		return ddmForm;
+	}
+
+}
